+++ docs/Changelog.md
@@ -11,7 +11,8 @@ Nussknacker versions
 1.2.0 (Not released yet)
 * [#2535](https://github.com/TouK/nussknacker/pull/2535) Rename `standalone` to `request-response`, 
   move request-response modules to `base` dir. Also - small refactorings in the engine and configuration format
-     
+* [#2483](https://github.com/TouK/nussknacker/pull/2483) Embedded DeploymentManager for Lite Streaming.
+
 1.1.0 (Not released yet)
 ------------------------
 * [#2176](https://github.com/TouK/nussknacker/pull/2176) Allow to enrich periodic scenario config on initial schedule and each deployment.